city_data_project/requirements.txt
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0
astral==3.2
geopy==2.4.1
